Updated solution of problem 10 from acmp.ru

+++ acmp.ru/001...100/010/10.cpp
@@ -7,9 +7,9 @@ int main()
 {
 	freopen("input.txt","r",stdin);
 	freopen("output.txt","w",stdout);
-	int a,b,c,d;
+	long long a,b,c,d;
 	cin>>a>>b>>c>>d;
-	for (int x=-100;x<=100;x++)
+	for (long long x=-100;x<=100;x++)
 		if (a*x*x*x+b*x*x+c*x+d==0)
 			cout<<x<<" ";
 	return 0;
